Covid infections
This is a simple program that uses an API to get recent data on Covid infections per 100K people, per country. I created this program because I couldn't find a resource that would display this metric in a simple and easily accessible manner.
This is a shortened and simplified version that can be used for further development.